# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      detail:
        description: 详细错误信息（可选）
        example: 详细错误信息
        type: string
      error:
        description: 错误信息
        example: 操作失败
        type: string
    type: object
...
host: localhost:58091
info:
  contact:
    email: im.hanxi@gmail.com
    name: API Support
    url: https://github.com/songloft-org/songloft
  description: 轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Songloft API
  version: 2.12.1
paths:
...
  /jsplugin/{entryPath}:
    get:
      description: JS 插件入口 HTML。{entryPath} 由运行时按已安装插件决定，注入 <base> 标签和 auth-bridge
        脚本后返回 static/index.html。无需认证。
      parameters:
      - description: 插件入口标识（运行时动态）
        in: path
        name: entryPath
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: 插件 index.html（已注入 <base> 与 auth-bridge）
          schema:
            type: string
        "404":
          description: 插件未安装或缺 static/index.html
          schema:
            type: string
      summary: 插件根页面（动态路由）
      tags:
      - JS 插件
  /jsplugin/{entryPath}/{path}:
    delete:
      consumes:
      - application/json
      description: 接受任意 HTTP 方法，分发到插件 static 兜底、入站 WebSocket upgrade（调用 onWebSocket），或转发到
        QuickJS 沙盒中的插件代码。{entryPath} 和子路径均由运行时决定，OpenAPI 仅作占位。需要 BearerAuth。
      parameters:
      - description: 插件入口标识（运行时动态）
        in: path
        name: entryPath
        required: true
        type: string
      - description: 可选：放宽本次插件调用的超时（毫秒），宿主 clamp 到 [30000, 300000]。慢端点（如 yt-dlp 展开歌单）用于避免默认
          30s 超时
        in: header
        name: X-Plugin-Timeout-Ms
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 插件自定义响应
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 插件未启用
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 插件不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 插件不可用或运行异常（健康检查会自愈）
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: JS 运行时调用超时
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件 API 转发 catch-all（动态路由）
      tags:
      - JS 插件
    get:
      consumes:
      - application/json
      description: 接受任意 HTTP 方法，分发到插件 static 兜底、入站 WebSocket upgrade（调用 onWebSocket），或转发到
        QuickJS 沙盒中的插件代码。{entryPath} 和子路径均由运行时决定，OpenAPI 仅作占位。需要 BearerAuth。
      parameters:
      - description: 插件入口标识（运行时动态）
        in: path
        name: entryPath
        required: true
        type: string
      - description: 可选：放宽本次插件调用的超时（毫秒），宿主 clamp 到 [30000, 300000]。慢端点（如 yt-dlp 展开歌单）用于避免默认
          30s 超时
        in: header
        name: X-Plugin-Timeout-Ms
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 插件自定义响应
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 插件未启用
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 插件不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 插件不可用或运行异常（健康检查会自愈）
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: JS 运行时调用超时
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件 API 转发 catch-all（动态路由）
      tags:
      - JS 插件
    post:
      consumes:
      - application/json
      description: 接受任意 HTTP 方法，分发到插件 static 兜底、入站 WebSocket upgrade（调用 onWebSocket），或转发到
        QuickJS 沙盒中的插件代码。{entryPath} 和子路径均由运行时决定，OpenAPI 仅作占位。需要 BearerAuth。
      parameters:
      - description: 插件入口标识（运行时动态）
        in: path
        name: entryPath
        required: true
        type: string
      - description: 可选：放宽本次插件调用的超时（毫秒），宿主 clamp 到 [30000, 300000]。慢端点（如 yt-dlp 展开歌单）用于避免默认
          30s 超时
        in: header
        name: X-Plugin-Timeout-Ms
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 插件自定义响应
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 插件未启用
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 插件不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 插件不可用或运行异常（健康检查会自愈）
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: JS 运行时调用超时
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件 API 转发 catch-all（动态路由）
      tags:
      - JS 插件
    put:
      consumes:
      - application/json
      description: 接受任意 HTTP 方法，分发到插件 static 兜底、入站 WebSocket upgrade（调用 onWebSocket），或转发到
        QuickJS 沙盒中的插件代码。{entryPath} 和子路径均由运行时决定，OpenAPI 仅作占位。需要 BearerAuth。
      parameters:
      - description: 插件入口标识（运行时动态）
        in: path
        name: entryPath
        required: true
        type: string
      - description: 可选：放宽本次插件调用的超时（毫秒），宿主 clamp 到 [30000, 300000]。慢端点（如 yt-dlp 展开歌单）用于避免默认
          30s 超时
        in: header
        name: X-Plugin-Timeout-Ms
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 插件自定义响应
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 插件未启用
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 插件不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 插件不可用或运行异常（健康检查会自愈）
          schema:
            additionalProperties:
              type: string
            type: object
        "504":
          description: JS 运行时调用超时
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 插件 API 转发 catch-all（动态路由）
      tags:
      - JS 插件
...
  /jsplugins:
    get:
      consumes:
      - application/json
      description: |-
        获取 JS 插件列表，形如 {"plugins": [jsplugin.JSPlugin, ...]}（响应类型是 map，
        单个插件对象的完整字段见 jsplugin.JSPlugin 定义）。
        其中 render_engine 是插件在自己 plugin.json 的 renderEngine 字段里声明的页面渲染引擎，
        取值 "webview"（系统 WebView）或 "lynx"（Lynx 原生渲染）；**空串表示跟随宿主默认**，客户端需自行映射为 webview。
      produces:
      - application/json
      responses:
        "200":
          description: JS插件列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 列出所有 JS 插件
      tags:
      - JS插件管理
  /jsplugins/{id}:
    delete:
      consumes:
      - application/json
      description: 根据插件ID删除 JS 插件。可通过 keep_data 参数保留插件数据目录（文件系统存储），持久化存储（数据库）始终保留。
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      - description: 是否保留插件数据目录（true/false，默认 false）
        in: query
        name: keep_data
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 插件不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除 JS 插件
      tags:
      - JS插件管理
    get:
      consumes:
      - application/json
      description: 根据插件ID获取 JS 插件的详细信息
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: JS插件信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 插件不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取 JS 插件详情
      tags:
      - JS插件管理
    put:
      consumes:
      - multipart/form-data
      description: 上传新的 JS 插件文件以更新现有插件
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      - description: JS插件文件 (.jsplugin.zip)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求数据错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 插件不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新 JS 插件
      tags:
      - JS插件管理
  /jsplugins/{id}/check-update:
    get:
      consumes:
      - application/json
      description: 检查指定 JS 插件的远程更新
      parameters:
      - description: 插件ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 更新信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 插件不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 检查 JS 插件更新
      tags:
      - JS插件管理
...
  /jsplugins/storage/cleanup:
    post:
      description: 删除 plugin_storage 表中不属于任何已安装插件的数据。当插件被卸载后，其持久化存储数据会保留在数据库中；此端点用于清理这些无主数据。
      produces:
      - application/json
      responses:
        "200":
          description: 清理完成
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 清理孤儿持久化存储
      tags:
      - JS插件管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer {token}" 进行认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"